+++ app/articles/[id]/page.tsx
@@ -1,4 +1,3 @@
-'use client';
 // TODO: 추후 API와 연동 시 작가ID/글 제목으로 라우팅 변경
 
 import SidePanelContainer from '@/app/components/SmartDrag/SidePanel/SidePanelContainer';
@@ -6,7 +5,10 @@ import ArticleContainer from '@/app/components/Article/ArticleContainer';
 import HeartIcon from '@/app/components/Icons/HeartIcon';
 import ShareIcon from '@/app/components/Icons/ShareIcon';
 
-export default function ArticlePage({ params }: { params: { id: string } }) {
+interface PropsType {
+  params: { id: string };
+}
+export default function ArticlePage({ params }: PropsType) {
   return (
     <div className="overflow-hidden flex flex-row justify-between h-full">
       {/* 좌측 좋아요 이동 바 */}
@@ -19,7 +21,7 @@ export default function ArticlePage({ params }: { params: { id: string } }) {
         </div>
       </div>
       {/* 본문 영역 */}
-      <ArticleContainer pageId={params.id} />
+      <ArticleContainer pageParams={params} />
       <SidePanelContainer />
     </div>
   );
