+++ app/components/Icons/HeartIcon.tsx
@@ -1,3 +1,4 @@
+'use client';
 import React, { useState } from 'react';
 
 interface PropsType {
